AI model application/backend/requirements.txt
# YOLOv8 Object Detection API Dependencies

# Flask web framework
flask==3.0.0
flask-cors==4.0.0

# YOLOv8 va computer vision
ultralytics==8.1.0
opencv-python==4.8.1.78

# Image processing
Pillow==10.1.0
numpy==1.24.3

# ONNX runtime (YOLOv8 optimizatsiya uchun)
onnxruntime==1.16.3

# Torch (YOLOv8 backend)
torch==2.1.0
torchvision==0.16.0

# Qo'shimcha utilities
requests==2.31.0
